requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
diffusers>=0.35.0
transformers>=4.52.0
accelerate>=0.33.0
safetensors>=0.4.0
pillow>=10.0.0
torch>=2.7.0

